Remove archive on any post-write failure; tidy downloader tests
Address review findings on the dependency-archive downloader.

Atomic cleanup on all post-write failures: `download_from_urls` and
`download_archive` now remove the archive through the same `cap_std`
capability whenever a failure occurs after it is written — a checksum
fetch/parse error, a reopen failure, the checksum mismatch, or an
`io::copy` write failure — via a shared `remove_partial_archive` helper
that logs (never masks) cleanup errors. A retry therefore never observes a
partial or unverified file. A boundary test now asserts the archive is
absent after a checksum-retrieval failure.

Boundary test tidy-up: `test_agent` becomes an rstest `#[fixture]` named
`agent`; a `download_harness` constructor builds the server and temp
destination from a route table (a plain fn, since rstest fixtures cannot
take per-call routes). The four workflow tests consume these; the two
non-UTF-8 tests stay separate. The trait/concrete-downloader imports used
only by the `#[cfg(unix)]` production-path test are now `#[cfg(unix)]`-gated.

Module hygiene: the loopback HTTP test server moves to a new
`http_test_server.rs` sibling so every source file stays at or below 400
lines (downloader 400, checksum 227, boundary tests 299, server 158).

Docs: the developers-guide control-flow list now credits `checksum.rs`
(`compute_sha256` streaming/rendering, `verify_archive_checksum`) and the
`downloader.rs` bullet describes only URL resolution, download/reopen, and
orchestration of the verification call.
